+++ src/main/groovy/io/xh/hoist/http/JSONClient.groovy
@@ -19,7 +19,7 @@ import org.apache.hc.client5.http.impl.classic.HttpClients
 import static io.opentelemetry.api.trace.SpanKind.CLIENT
 import static io.xh.hoist.telemetry.ObservedRun.observe
 import static io.xh.hoist.util.StringUtils.elide
-import static io.xh.hoist.util.Utils.traceService
+import static io.xh.hoist.util.Utils.traceSupportService
 import static org.apache.hc.core5.http.HttpStatus.SC_NO_CONTENT
 import static org.apache.hc.core5.http.HttpStatus.SC_OK
 
@@ -159,7 +159,7 @@ class JSONClient {
                 ]
             )
             .run { SpanRef span ->
-                traceService.injectContext(method)
+                traceSupportService.injectContext(method)
                 def ret = client.execute(method)
                 span.setHttpStatus(ret.code)
                 ret