+++ src/main/groovy/io/xh/hoist/telemetry/MetricsConfig.groovy
@@ -6,17 +6,17 @@
  */
 package io.xh.hoist.telemetry
 
-import groovy.transform.InheritConstructors
 import io.xh.hoist.config.TypedConfigMap
 
 /**
  * Typed representation of `xhMetricsConfig` values.
  */
-@InheritConstructors
 class MetricsConfig extends TypedConfigMap {
     String namespace
     boolean prometheusEnabled
     Map prometheusConfig
     boolean otlpEnabled
     Map otlpConfig
+
+    MetricsConfig(Map args) { init(args) }
 }