+++ src/main/groovy/io/xh/hoist/telemetry/TraceConfig.groovy
@@ -6,13 +6,11 @@
  */
 package io.xh.hoist.telemetry
 
-import groovy.transform.InheritConstructors
 import io.xh.hoist.config.TypedConfigMap
 
 /**
  * Typed representation of `xhTraceConfig` values.
  */
-@InheritConstructors
 class TraceConfig extends TypedConfigMap {
     boolean enabled
     double sampleRate
@@ -24,4 +22,9 @@ class TraceConfig extends TypedConfigMap {
 
     /** Always export error spans, bypassing sample-rate filtering. Defaults to true. */
     boolean alwaysSampleErrors = true
+
+    /** Emit CLIENT spans for all JDBC operations. Defaults to false. */
+    boolean jdbcTracingEnabled = false
+
+    TraceConfig(args) { init(args) }
 }