+++ src/main/groovy/io/xh/hoist/util/Utils.groovy
@@ -30,6 +30,7 @@ import io.xh.hoist.security.BaseAuthenticationService
 import io.xh.hoist.user.BaseUserService
 import io.xh.hoist.user.IdentityService
 import io.xh.hoist.telemetry.TraceService
+import io.xh.hoist.telemetry.impl.TraceSupportService
 import io.xh.hoist.websocket.WebSocketService
 import org.grails.web.servlet.mvc.GrailsWebRequest
 
@@ -145,6 +146,10 @@ class Utils {
 
     static TraceService getTraceService() {
         return (TraceService) appContext.traceService
+    }
+
+    static TraceSupportService getTraceSupportService() {
+        return (TraceSupportService) appContext.traceSupportService
     }
 
     static WebSocketService getWebSocketService() {